feat: Edit View 레이아웃 수정, collectionView cell 생성
