+++ OpenMarket/OpenMarket/Scene/Edit/ProductsHorizontalCell.swift
@@ -1,0 +1,44 @@
+//
+//  HorizontalCell.swift
+//  OpenMarket
+//
+//  Created by 박세리 on 2022/05/24.
+//
+
+import UIKit
+final class ProductsHorizontalCell: UICollectionViewCell, BaseCell {
+    static var identifier: String {
+        return String(describing: self)
+    }
+    
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        addSubviews()
+        makeConstraints()
+    }
+    
+    required init?(coder: NSCoder) {
+        super.init(coder: coder)
+        addSubviews()
+        makeConstraints()
+    }
+    
+    let productImageView: UIImageView = {
+        let imageView = UIImageView()
+        imageView.translatesAutoresizingMaskIntoConstraints = false
+        return imageView
+    }()
+    
+    func addSubviews() {
+        self.addSubview(productImageView)
+    }
+    
+    func makeConstraints() {
+        NSLayoutConstraint.activate([
+            productImageView.topAnchor.constraint(equalTo: contentView.topAnchor),
+            productImageView.leadingAnchor.constraint(equalTo: contentView.leadingAnchor),
+            productImageView.trailingAnchor.constraint(equalTo: contentView.trailingAnchor),
+            productImageView.bottomAnchor.constraint(equalTo: contentView.bottomAnchor)
+        ])
+    }
+}
